refactor: convert rankings.js to ts
